Move contexts to separate directory and remove them from .eslintignore.
