+++ src/contexts/SortContext.js
@@ -1,6 +1,5 @@
-/* Encapsulates the as-yet-unmanaged sorting behavior for cube lists. */
-
-import React, { Component } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 const SortContext = React.createContext({
   primary: 'Color Category',
@@ -11,10 +10,13 @@ const SortContext = React.createContext({
 export class SortContextProvider extends React.Component {
   constructor(props) {
     super(props);
+    const {
+      defaultSorts: [primary = 'Color Category', secondary = 'Types-Multicolor'],
+    } = this.props;
 
     this.state = {
-      primary: this.props.defaultSorts[0] || 'Color Category',
-      secondary: this.props.defaultSorts[1] || 'Types-Multicolor',
+      primary,
+      secondary,
       tertiary: 'CMC2',
     };
 
@@ -24,12 +26,13 @@ export class SortContextProvider extends React.Component {
   componentDidMount() {
     for (const stage of ['primary', 'secondary']) {
       const select = document.getElementById(`${stage}SortSelect`);
-      if (!select) continue;
-      select.addEventListener('change', (event) => {
-        this.setState({
-          [stage]: event.target.value,
+      if (select) {
+        select.addEventListener('change', (event) => {
+          this.setState({
+            [stage]: event.target.value,
+          });
         });
-      });
+      }
     }
   }
 
@@ -45,6 +48,10 @@ export class SortContextProvider extends React.Component {
     return <SortContext.Provider value={value} {...this.props} />;
   }
 }
+
+SortContextProvider.propTypes = {
+  defaultSorts: PropTypes.arrayOf(PropTypes.string).isRequired,
+};
 
 SortContext.Wrapped = (Tag) => (props) => (
   <SortContext.Consumer>{(value) => <Tag {...props} {...value} />}</SortContext.Consumer>
