+++ src/pages/CubeListPage.js
@@ -6,21 +6,21 @@ import LocalStorage from 'utils/LocalStorage';
 import Query from 'utils/Query';
 
 import CardModalForm from 'components/CardModalForm';
-import { ChangelistContextProvider } from 'components/ChangelistContext';
+import { ChangelistContextProvider } from 'contexts/ChangelistContext';
 import ClientOnly from 'components/ClientOnly';
-import CubeContext from 'components/CubeContext';
+import CubeContext from 'contexts/CubeContext';
 import CubeListNavbar from 'components/CubeListNavbar';
 import CurveView from 'components/CurveView';
-import DisplayContext, { DisplayContextProvider } from 'components/DisplayContext';
+import DisplayContext, { DisplayContextProvider } from 'contexts/DisplayContext';
 import DynamicFlash from 'components/DynamicFlash';
 import ErrorBoundary from 'components/ErrorBoundary';
 import GroupModal from 'components/GroupModal';
 import ListView from 'components/ListView';
 import Maybeboard from 'components/Maybeboard';
-import { MaybeboardContextProvider } from 'components/MaybeboardContext';
-import { SortContextProvider } from 'components/SortContext';
+import { MaybeboardContextProvider } from 'contexts/MaybeboardContext';
+import { SortContextProvider } from 'contexts/SortContext';
 import TableView from 'components/TableView';
-import { TAG_COLORS, TagContextProvider } from 'components/TagContext';
+import { TAG_COLORS, TagContextProvider } from 'contexts/TagContext';
 import VisualSpoiler from 'components/VisualSpoiler';
 import CubeLayout from 'layouts/CubeLayout';
 import MainLayout from 'layouts/MainLayout';
